Refactored to use a utilities class
