+++ CoreDataAccessView/test/unit/src/au/gov/asd/tac/constellation/views/dataaccess/panes/DataAccessSearchProviderNGTest.java
@@ -17,6 +17,7 @@ package au.gov.asd.tac.constellation.views.dataaccess.panes;
 
 import au.gov.asd.tac.constellation.utilities.color.ConstellationColor;
 import au.gov.asd.tac.constellation.utilities.icon.UserInterfaceIconProvider;
+import au.gov.asd.tac.constellation.utilities.qs.QuickSearchUtilities;
 import au.gov.asd.tac.constellation.views.dataaccess.components.DataAccessTabPane;
 import au.gov.asd.tac.constellation.views.dataaccess.tasks.ShowDataAccessPluginTask;
 import au.gov.asd.tac.constellation.views.dataaccess.utilities.DataAccessUtilities;
@@ -178,15 +179,15 @@ public class DataAccessSearchProviderNGTest {
         when(request.getText()).thenReturn("Select");
 
         // Return a valid response when results are added that match the expected
-        when(response.addResult(Mockito.any(), Mockito.eq(DataAccessSearchProvider.CIRCLED_D + "  Select Top N"))).thenReturn(true);
-        when(response.addResult(Mockito.any(), Mockito.eq(DataAccessSearchProvider.CIRCLED_D + "  Select All"))).thenReturn(true);
+        when(response.addResult(Mockito.any(), Mockito.eq(QuickSearchUtilities.CIRCLED_D + "  Select Top N"))).thenReturn(true);
+        when(response.addResult(Mockito.any(), Mockito.eq(QuickSearchUtilities.CIRCLED_D + "  Select All"))).thenReturn(true);
 
         DataAccessSearchProvider instance = new DataAccessSearchProvider();
         instance.evaluate(request, response);
 
         // Verify that addResult was called on the correct plugins
-        verify(response, times(1)).addResult(Mockito.any(), Mockito.eq(DataAccessSearchProvider.CIRCLED_D + "  Select Top N"));
-        verify(response, times(1)).addResult(Mockito.any(), Mockito.eq(DataAccessSearchProvider.CIRCLED_D + "  Select All"));
+        verify(response, times(1)).addResult(Mockito.any(), Mockito.eq(QuickSearchUtilities.CIRCLED_D + "  Select Top N"));
+        verify(response, times(1)).addResult(Mockito.any(), Mockito.eq(QuickSearchUtilities.CIRCLED_D + "  Select All"));
     }
 
     ////////////////////////////////////////////////////////////////////////////
